requirements.txt
anyio==3.6.2
APScheduler==3.6.3
backports.zoneinfo==0.2.1;python_version<"3.9"
blinker==1.6.2
cachetools==4.2.2
certifi==2023.7.22
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
h11==0.14.0
httpcore>=1.0.0
httpx>=0.26.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyaes==1.6.1
pyasn1==0.5.0
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
Telethon==1.28.5
tornado==6.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
Werkzeug==2.3.4
zipp==3.15.0
